// apps/game-client/src/lib/api/generated/main/model/member-summary-response-dto-output.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */

export interface MemberSummaryResponseDtoOutput {
  id: number;
  userId: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  color?: number | null;
}